[WIP] Convert MemoPlaintext to String type
We need a fixed length of 512 bytes for passing the memo plaintext
to cryptographic functions. Everywhere else, though, a String is most
convenient. Here we add some conversion traits to make working
with MemoPlaintext a bit more ergonomic. Ideally we'd perform validation
such that a too-long String cannot be used to construct a MemoPlaintext.